parser/requirements.txt
pdfplumber>=0.10.0
camelot-py[cv]>=0.11.0
rapidfuzz>=3.0.0
pydantic>=2.0.0
PyYAML>=6.0
typer>=0.9.0
rich>=13.0.0
pytest>=7.0.0
pillow>=10.0.0